refactor: 회원 정보 요청 애노테이션 네이밍 및 패키지 변경
