+++ space-d/src/main/java/com/dnd/spaced/core/account/presentation/AccountController.java
@@ -4,7 +4,7 @@ import com.dnd.spaced.core.account.application.AccountService;
 import com.dnd.spaced.core.account.application.dto.request.ChangeCareerInfoRequest;
 import com.dnd.spaced.core.account.application.dto.request.ChangeProfileInfoRequest;
 import com.dnd.spaced.core.account.application.dto.response.AccountResponse;
-import com.dnd.spaced.global.auth.AuthAccount;
+import com.dnd.spaced.global.auth.resolver.CurrentAccountInfo;
 import com.dnd.spaced.global.auth.resolver.AuthAccountInfo;
 import com.dnd.spaced.global.consts.controller.ResponseEntityConst;
 import jakarta.validation.Valid;
@@ -25,7 +25,7 @@ public class AccountController {
     private final AccountService accountService;
 
     @DeleteMapping("/withdrawal")
-    public ResponseEntity<Void> withdrawal(@AuthAccount AuthAccountInfo accountInfo) {
+    public ResponseEntity<Void> withdrawal(@CurrentAccountInfo AuthAccountInfo accountInfo) {
         accountService.withdrawal(accountInfo.accountId());
 
         return ResponseEntityConst.NO_CONTENT;
@@ -33,7 +33,7 @@ public class AccountController {
 
     @PutMapping("/career-info")
     public ResponseEntity<Void> changeCareerInfo(
-            @AuthAccount AuthAccountInfo accountInfo,
+            @CurrentAccountInfo AuthAccountInfo accountInfo,
             @Valid @RequestBody ChangeCareerInfoRequest request
     ) {
         accountService.changeCareerInfo(accountInfo.accountId(), request);
@@ -43,7 +43,7 @@ public class AccountController {
 
     @PutMapping("/profile-info")
     public ResponseEntity<Void> changeProfileInfo(
-            @AuthAccount AuthAccountInfo accountInfo,
+            @CurrentAccountInfo AuthAccountInfo accountInfo,
             @Valid @RequestBody ChangeProfileInfoRequest request
     ) {
         accountService.changeProfileInfo(accountInfo.accountId(), request);
@@ -52,7 +52,7 @@ public class AccountController {
     }
 
     @GetMapping
-    public ResponseEntity<AccountResponse> readAccount(@AuthAccount AuthAccountInfo accountInfo) {
+    public ResponseEntity<AccountResponse> readAccount(@CurrentAccountInfo AuthAccountInfo accountInfo) {
         AccountResponse response = accountService.readAccount(accountInfo.accountId());
 
         return ResponseEntity.ok(response);
