+++ space-d/src/main/java/com/dnd/spaced/core/bookmark/presentation/BookmarkController.java
@@ -4,7 +4,7 @@ import com.dnd.spaced.core.bookmark.application.BookmarkService;
 import com.dnd.spaced.core.bookmark.application.dto.request.CreateBookmarkRequest;
 import com.dnd.spaced.core.bookmark.application.dto.request.ReadAllBookmarkRequest;
 import com.dnd.spaced.core.bookmark.application.dto.response.BookmarkCollectionResponse;
-import com.dnd.spaced.global.auth.AuthAccount;
+import com.dnd.spaced.global.auth.resolver.CurrentAccountInfo;
 import com.dnd.spaced.global.auth.resolver.AuthAccountInfo;
 import com.dnd.spaced.global.consts.controller.ResponseEntityConst;
 import com.dnd.spaced.global.resolver.bookmark.BookmarkPageable;
@@ -29,7 +29,7 @@ public class BookmarkController {
 
     @GetMapping
     public ResponseEntity<BookmarkCollectionResponse> readBookmarks(
-            @AuthAccount AuthAccountInfo accountInfo,
+            @CurrentAccountInfo AuthAccountInfo accountInfo,
             ReadAllBookmarkRequest request,
             @BookmarkPageable Pageable pageable
     ) {
@@ -40,7 +40,7 @@ public class BookmarkController {
 
     @PostMapping
     public ResponseEntity<Void> createBookmark(
-            @AuthAccount AuthAccountInfo accountInfo,
+            @CurrentAccountInfo AuthAccountInfo accountInfo,
             @RequestBody @Valid CreateBookmarkRequest request
     ) {
         bookmarkService.createBookmark(accountInfo.accountId(), request);
@@ -50,7 +50,7 @@ public class BookmarkController {
 
     @DeleteMapping("{bookmarkId}")
     public ResponseEntity<Void> deleteBookmark(
-            @AuthAccount AuthAccountInfo accountInfo,
+            @CurrentAccountInfo AuthAccountInfo accountInfo,
             @PathVariable Long bookmarkId
     ) {
         bookmarkService.deleteBookmark(accountInfo.accountId(), bookmarkId);
