+++ space-d/src/main/java/com/dnd/spaced/core/like/presentation/LikeController.java
@@ -1,7 +1,7 @@
 package com.dnd.spaced.core.like.presentation;
 
 import com.dnd.spaced.core.like.application.LikeService;
-import com.dnd.spaced.global.auth.AuthAccount;
+import com.dnd.spaced.global.auth.resolver.CurrentAccountInfo;
 import com.dnd.spaced.global.auth.resolver.AuthAccountInfo;
 import com.dnd.spaced.global.consts.controller.ResponseEntityConst;
 import lombok.RequiredArgsConstructor;
@@ -19,7 +19,7 @@ public class LikeController {
     private final LikeService likeService;
 
     @PostMapping
-    public ResponseEntity<Void> processLike(@AuthAccount AuthAccountInfo accountInfo, @PathVariable Long commentId) {
+    public ResponseEntity<Void> processLike(@CurrentAccountInfo AuthAccountInfo accountInfo, @PathVariable Long commentId) {
         likeService.processLike(accountInfo.accountId(), commentId);
 
         return ResponseEntityConst.NO_CONTENT;
