+++ space-d/src/main/java/com/dnd/spaced/global/auth/resolver/CurrentAccountInfo.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.global.auth;
+package com.dnd.spaced.global.auth.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,7 +7,7 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface AuthAccount {
+public @interface CurrentAccountInfo {
 
     boolean required() default true;
 }
